firstCommit
Models and training data are not on Github due to the size constraints. Disclaimer: The model is built from scratch with only one GPU available so performance-wise it is not awesome. It shows great improvement after training it to make the cross-entropy loss down from ~9 ish to 3ish. I followed instructors to build this for learning purpose.